Moved ParsingUtils to utils
