+++ src/main/java/org/jetbrains/bsp/bazel/server/utils/BUILD
@@ -4,4 +4,11 @@ java_library(
     name = "utils",
     srcs = glob(["*.java"]),
     visibility = ["//visibility:public"],
+    deps = [
+        "//src/main/java/org/jetbrains/bsp/bazel/common",
+        "@maven//:ch_epfl_scala_bsp4j",
+        "@maven//:com_google_guava_guava",
+        "@maven//:org_apache_logging_log4j_log4j_api",
+        "@maven//:org_apache_logging_log4j_log4j_core",
+    ],
 )
